requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
Pillow>=10.0.0
django-allauth>=0.62.0
xhtml2pdf>=0.2.15